requirements.txt
lightning==2.0.3
matplotlib==3.7.1
Pillow==9.5.0
Pillow==10.0.1
setuptools==65.5.0
torch==2.0.1
torchvision==0.15.2
